added action modules and ported 2 functions to basic actions
